Add dataAccount object
